requirements.txt
Flask
Flask-RESTful
flask-restx
flask-accepts
numpy
opencv-python
image
python-telegram-bot

